as.bootstrap.scores

Returns a dataframe with all the bootstrap score in a TRONCO model. It is possible to specify a subset of events or models if multiple reconstruction have been performed.

as.bootstrap.scores(x, events = as.events(x), models = names(x$model))

Arguments

x
A TRONCO model.

events
A subset of events as of as.events(x), all by default.

models
A subset of reconstructed models, all by default.

Value

All the bootstrap scores in a TRONCO model

Examples

data(test_model)
as.bootstrap.scores(test_model)

#> $capri_bic
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2            missense_point_mutations KIT
#> 2                        ins_del ASXL1           missense_point_mutations TET2
#> 3      missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 4  missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 5  missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 6        missense_point_mutations TET2                            ins_del EZH2
#> 7        missense_point_mutations EZH2                            ins_del TET2
#> 8        missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 9        missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 10        missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 11              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 12              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 13      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 14      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#>    OBS.SELECTS OBS.SELECTED
#> 1            3            2
#> 2            7            3
#> 3           14            4
#> 4            3            1
#> 5            3            1
#> 6            3            1
#> 7            6            3
#> 8            6            1
#> 9            6            1
#> 10           2            1
#> 11           2            1
#> 12           2            1
#> 13           4            3
#> 14           4            1
#> 
#> $capri_aic
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2     missense_point_mutations CBL_Ex_8_9
#> 2                         ins_del TET2            missense_point_mutations KIT
#> 3                        ins_del ASXL1     missense_point_mutations KRAS_Ex2_3
#> 4                        ins_del ASXL1           missense_point_mutations TET2
#> 5      missense_point_mutations SETBP1           nonsense_point_mutations EZH2
#> 6      missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 7  missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 8  missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 9        missense_point_mutations TET2                            ins_del EZH2
#> 10       missense_point_mutations TET2                  nonsense_ins_del CEBPA
#> 11       missense_point_mutations EZH2                            ins_del TET2
#> 12       missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 13       missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 14      missense_point_mutations CSF3R           missense_point_mutations TET2
#> 15        missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 16              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 17              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 18      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 19      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#> 20                    Pattern OR_CSF3R                  nonsense_ins_del CEBPA
#>    OBS.SELECTS OBS.SELECTED
#> 1            3            3
#> 2            3            2
#> 3            7            1
#> 4            7            3
#> 5           14            1
#> 6           14            4
#> 7            3            1
#> 8            3            1
#> 9            3            1
#> 10           3            2
#> 11           6            3
#> 12           6            1
#> 13           6            1
#> 14           4            3
#> 15           2            1
#> 16           2            1
#> 17           2            1
#> 18           4            3
#> 19           4            1
#> 20           6            2
#> 

as.bootstrap.scores(test_model, events=as.events(test_model)[5:15,])

#> $capri_bic
#>                               SELECTS                                SELECTED
#> 1 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 2       missense_point_mutations EZH2          missense_point_mutations CEBPA
#>   OBS.SELECTS OBS.SELECTED
#> 1           3            1
#> 2           6            1
#> 
#> $capri_aic
#>                               SELECTS                                SELECTED
#> 1 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 2       missense_point_mutations EZH2          missense_point_mutations CEBPA
#>   OBS.SELECTS OBS.SELECTED
#> 1           3            1
#> 2           6            1
#>