as.conditional.probs

Extract the conditional probabilities from a TRONCO model. The return matrix is indexed with rownames which represent genotype keys - these can be resolved with function keysToNames. It is possible to specify a subset of events to build the matrix, a subset of models if multiple reconstruction have been performed. Also, either the observed or fit probabilities can be extracted.

as.conditional.probs(
  x,
  events = as.events(x),
  models = names(x$model),
  type = "observed"
)

Arguments

x

A TRONCO model.

events

A subset of events as of as.events(x), all by default.

models

A subset of reconstructed models, all by default.

type

observed ('observed')

Value

The conditional probabilities in a TRONCO model.

Details

#' @examples data(test_model) as.conditional.probs(test_model) as.conditional.probs(test_model, events=as.events(test_model)[5:15,])