as.confidence

Return confidence information for a TRONCO model. Available information are: temporal priority (tp), probability raising (pr), hypergeometric test (hg), parametric (pb), non parametric (npb) or statistical (sb) bootstrap, entropy loss (eloss), prediction error (prederr). Confidence is available only once a model has been reconstructed with any of the algorithms implemented in TRONCO. If more than one model has been reconstructed - for instance via multiple regularizations - confidence information is appropriately nested. The requested confidence is specified via vector parameter conf.

as.confidence(x, conf, models = names(x$model))

Arguments

x

A TRONCO model.

conf

A vector with any of 'tp', 'pr', 'hg', 'npb', 'pb', 'sb', 'eloss', 'prederr' or 'posterr'.

models

The name of the models to extract, all by default.

Value

A list of matrices with the event-to-event confidence.

Examples

data(test_model)
as.confidence(test_model, conf='tp')
#> $tp
#>                gene 4       gene 5       gene 7 gene 29      gene 30
#> gene 4   1.000000e+00 1.455680e-26 1.000000e+00       1 9.251274e-01
#> gene 5   1.000000e+00 1.000000e+00 1.000000e+00       1 1.000000e+00
#> gene 7   5.329851e-38 3.926171e-55 1.000000e+00       1 1.724654e-36
#> gene 29  4.736819e-56 9.709084e-57 6.798626e-51       1 4.357290e-56
#> gene 30  7.503856e-02 2.295716e-31 1.000000e+00       1 1.000000e+00
#> gene 31  1.000000e+00 4.509268e-01 1.000000e+00       1 1.000000e+00
#> gene 32  5.489782e-01 1.282066e-25 1.000000e+00       1 9.344646e-01
#> gene 33  9.373014e-28 1.918776e-52 9.991563e-01       1 3.120373e-25
#> gene 34  7.643872e-01 1.464270e-25 1.000000e+00       1 9.860117e-01
#> gene 36  1.000000e+00 1.002462e-01 1.000000e+00       1 1.000000e+00
#> gene 40  1.000000e+00 7.909232e-01 1.000000e+00       1 1.000000e+00
#> gene 44  1.000000e+00 7.337264e-01 1.000000e+00       1 1.000000e+00
#> gene 47  1.000000e+00 3.174193e-01 1.000000e+00       1 1.000000e+00
#> gene 49  1.000000e+00 7.387701e-01 1.000000e+00       1 1.000000e+00
#> gene 50  1.000000e+00 1.779009e-01 1.000000e+00       1 1.000000e+00
#> gene 51  1.000000e+00 3.036831e-01 1.000000e+00       1 1.000000e+00
#> gene 52  1.000000e+00 1.593646e-01 1.000000e+00       1 1.000000e+00
#> gene 53  1.000000e+00 5.479794e-02 1.000000e+00       1 1.000000e+00
#> gene 54  1.000000e+00 4.507570e-01 1.000000e+00       1 1.000000e+00
#> gene 55  8.972807e-11 8.857734e-45 1.000000e+00       1 5.160503e-08
#> gene 56  9.999986e-01 8.703374e-13 1.000000e+00       1 1.000000e+00
#> gene 66  1.000000e+00 6.386084e-01 1.000000e+00       1 1.000000e+00
#> gene 69  1.000000e+00 6.115105e-01 1.000000e+00       1 1.000000e+00
#> gene 77  9.999998e-01 1.552462e-12 1.000000e+00       1 1.000000e+00
#> gene 88  1.126419e-18 2.950660e-49 1.000000e+00       1 9.388325e-16
#> gene 89  1.000000e+00 6.967985e-01 1.000000e+00       1 1.000000e+00
#> gene 91  8.337071e-07 1.622532e-41 1.000000e+00       1 2.148264e-04
#> gene 111 1.253088e-01 3.862890e-31 1.000000e+00       1 6.427562e-01
#> XOR_EZH2 1.657722e-44 1.000000e+00 3.787203e-05       1 1.100280e-43
#> OR_CSF3R 9.610599e-32 1.261960e-52 9.841773e-01       1 6.816777e-30
#>               gene 31      gene 32      gene 33      gene 34      gene 36
#> gene 4   1.547314e-26 4.514859e-01 1.000000e+00 2.359743e-01 2.342545e-22
#> gene 5   5.495566e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.999670e-01
#> gene 7   4.356112e-55 2.278343e-38 8.470797e-04 4.899826e-40 2.294724e-54
#> gene 29  1.064251e-56 3.984755e-56 6.509270e-52 4.145089e-56 1.447628e-56
#> gene 30  1.985077e-31 6.568508e-02 1.000000e+00 1.403023e-02 2.436703e-27
#> gene 31  1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 8.809486e-01
#> gene 32  1.764283e-25 1.000000e+00 1.000000e+00 2.823813e-01 1.551439e-21
#> gene 33  1.847050e-52 4.563330e-28 1.000000e+00 3.690851e-30 6.286873e-51
#> gene 34  2.002835e-25 7.180153e-01 1.000000e+00 1.000000e+00 3.799642e-21
#> gene 36  1.192927e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 40  8.227594e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.813282e-01
#> gene 44  7.662783e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.661370e-01
#> gene 47  3.798879e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.121589e-01
#> gene 49  7.806570e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.737212e-01
#> gene 50  2.095278e-01 1.000000e+00 1.000000e+00 1.000000e+00 6.505522e-01
#> gene 51  3.565901e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.052310e-01
#> gene 52  1.911901e-01 1.000000e+00 1.000000e+00 1.000000e+00 6.442144e-01
#> gene 53  7.289301e-02 1.000000e+00 1.000000e+00 1.000000e+00 4.286338e-01
#> gene 54  4.948992e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.682621e-01
#> gene 55  8.353704e-45 5.618501e-11 1.000000e+00 4.545591e-13 7.177829e-42
#> gene 56  1.321968e-12 9.999966e-01 1.000000e+00 9.999781e-01 2.419613e-09
#> gene 66  6.848490e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.484744e-01
#> gene 69  6.545899e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.397323e-01
#> gene 77  2.963835e-12 9.999994e-01 1.000000e+00 9.999967e-01 6.503131e-09
#> gene 88  2.781146e-49 7.497944e-19 9.999184e-01 3.539068e-21 4.251280e-47
#> gene 89  7.390193e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.635395e-01
#> gene 91  1.150112e-41 6.597718e-07 1.000000e+00 1.650994e-08 6.652005e-38
#> gene 111 2.698640e-31 1.144128e-01 1.000000e+00 2.935592e-02 5.982824e-27
#> XOR_EZH2 5.535303e-56 5.693025e-45 1.000000e+00 4.306813e-46 1.354017e-55
#> OR_CSF3R 1.483475e-52 4.947292e-32 1.396996e-01 6.632084e-34 1.866588e-51
#>               gene 40      gene 44      gene 47      gene 49      gene 50
#> gene 4   2.165386e-29 5.008267e-27 2.263522e-26 3.192407e-29 5.277778e-24
#> gene 5   2.094293e-01 2.666752e-01 6.830185e-01 2.616288e-01 8.224158e-01
#> gene 7   9.042160e-56 4.931317e-55 4.003732e-55 1.159402e-55 1.210707e-54
#> gene 29  8.419176e-57 9.175701e-57 6.888185e-57 8.336787e-57 1.287253e-56
#> gene 30  2.788036e-34 1.056706e-31 5.157664e-31 5.227785e-34 5.474974e-29
#> gene 31  1.775588e-01 2.340961e-01 6.205786e-01 2.197051e-01 7.908218e-01
#> gene 32  2.086112e-28 3.758638e-26 3.812068e-25 4.301145e-28 4.958266e-23
#> gene 33  1.214608e-53 2.709216e-52 2.057515e-52 1.677847e-53 1.443304e-51
#> gene 34  1.738248e-28 4.834981e-26 3.011753e-25 3.060859e-28 8.616538e-23
#> gene 36  1.872742e-02 3.395435e-02 1.881696e-01 2.635303e-02 3.498953e-01
#> gene 40  1.000000e+00 5.643395e-01 9.014077e-01 5.666406e-01 9.577996e-01
#> gene 44  4.361437e-01 1.000000e+00 8.638005e-01 5.024467e-01 9.314922e-01
#> gene 47  9.880598e-02 1.364671e-01 1.000000e+00 1.324596e-01 7.019801e-01
#> gene 49  4.338431e-01 4.980426e-01 8.678041e-01 1.000000e+00 9.435145e-01
#> gene 50  4.231026e-02 6.866853e-02 2.984434e-01 5.662415e-02 1.000000e+00
#> gene 51  9.189371e-02 1.329884e-01 4.624964e-01 1.189435e-01 6.829473e-01
#> gene 52  3.484817e-02 5.951470e-02 2.686340e-01 4.657274e-02 4.859700e-01
#> gene 53  7.824031e-03 1.625413e-02 1.062676e-01 1.103062e-02 2.761876e-01
#> gene 54  1.806809e-01 2.360790e-01 6.260883e-01 2.254798e-01 7.775074e-01
#> gene 55  5.230203e-47 1.126160e-44 1.553198e-44 9.837351e-47 4.491146e-43
#> gene 56  3.754344e-15 1.360701e-13 2.654492e-12 6.689968e-15 1.688733e-10
#> gene 66  3.300539e-01 3.940738e-01 7.915445e-01 3.900800e-01 8.992429e-01
#> gene 69  3.072810e-01 3.754195e-01 7.592675e-01 3.602876e-01 8.827444e-01
#> gene 77  6.488969e-15 2.216224e-13 4.511947e-12 1.134950e-14 4.576681e-10
#> gene 88  5.425064e-51 4.217177e-49 4.953952e-49 9.833821e-51 5.314708e-48
#> gene 89  3.852445e-01 4.500171e-01 8.407157e-01 4.515920e-01 9.251892e-01
#> gene 91  3.474747e-44 1.925081e-41 2.180308e-41 5.755818e-44 1.661066e-39
#> gene 111 3.779886e-34 2.032201e-31 5.878879e-31 5.866015e-34 9.465023e-29
#> XOR_EZH2 2.021003e-56 5.439267e-56 4.632102e-56 2.432208e-56 9.923625e-56
#> OR_CSF3R 1.546387e-53 1.361436e-52 1.636824e-52 2.239969e-53 7.180214e-52
#>               gene 51      gene 52      gene 53      gene 54      gene 55
#> gene 4   1.086371e-26 4.166290e-25 4.513530e-24 3.046136e-25 1.000000e+00
#> gene 5   6.967445e-01 8.409308e-01 9.453371e-01 5.497257e-01 1.000000e+00
#> gene 7   2.633366e-55 3.104739e-55 4.347804e-55 1.018560e-54 1.883964e-22
#> gene 29  9.352671e-57 1.181415e-56 1.069432e-56 1.136122e-56 6.828657e-56
#> gene 30  1.128523e-31 2.456006e-30 2.827924e-29 5.478567e-30 9.999999e-01
#> gene 31  6.438652e-01 8.091411e-01 9.272754e-01 5.055866e-01 1.000000e+00
#> gene 32  1.712602e-25 4.817431e-24 7.261810e-23 2.539887e-24 1.000000e+00
#> gene 33  8.126240e-53 1.574250e-52 3.167398e-52 1.025199e-51 1.663992e-11
#> gene 34  1.595057e-25 7.199230e-24 1.048880e-22 3.874618e-24 1.000000e+00
#> gene 36  1.951038e-01 3.562367e-01 5.718416e-01 1.319965e-01 1.000000e+00
#> gene 40  9.083082e-01 9.652457e-01 9.922022e-01 8.196405e-01 1.000000e+00
#> gene 44  8.672738e-01 9.406292e-01 9.837951e-01 7.642969e-01 1.000000e+00
#> gene 47  5.379924e-01 7.317696e-01 8.939573e-01 3.743749e-01 1.000000e+00
#> gene 49  8.813001e-01 9.535461e-01 9.890048e-01 7.748872e-01 1.000000e+00
#> gene 50  3.174865e-01 5.145136e-01 7.242182e-01 2.228543e-01 1.000000e+00
#> gene 51  1.000000e+00 7.069561e-01 8.729173e-01 3.628007e-01 1.000000e+00
#> gene 52  2.934635e-01 1.000000e+00 7.143980e-01 2.063617e-01 1.000000e+00
#> gene 53  1.273372e-01 2.860155e-01 1.000000e+00 8.201938e-02 1.000000e+00
#> gene 54  6.376563e-01 7.939845e-01 9.181644e-01 1.000000e+00 1.000000e+00
#> gene 55  2.928759e-45 1.450209e-44 7.671125e-44 1.599977e-43 1.000000e+00
#> gene 56  2.391171e-12 5.531515e-11 6.047829e-10 6.903869e-12 1.000000e+00
#> gene 66  8.078831e-01 9.129183e-01 9.744163e-01 6.818311e-01 1.000000e+00
#> gene 69  7.769980e-01 8.953302e-01 9.656875e-01 6.580630e-01 1.000000e+00
#> gene 77  5.924895e-12 1.791474e-10 2.121713e-09 1.367533e-11 1.000000e+00
#> gene 88  1.083763e-49 2.840822e-49 1.005260e-48 3.058781e-48 5.771054e-04
#> gene 89  8.516760e-01 9.368002e-01 9.836354e-01 7.347714e-01 1.000000e+00
#> gene 91  3.268978e-42 4.019945e-41 2.577904e-40 4.730861e-40 9.535215e-01
#> gene 111 1.201388e-31 3.564199e-30 3.490003e-29 1.059821e-29 1.000000e+00
#> XOR_EZH2 3.987969e-56 4.161263e-56 4.919078e-56 8.725395e-56 1.255183e-32
#> OR_CSF3R 9.498741e-53 1.490719e-52 2.977122e-52 4.358181e-52 1.000000e+00
#>               gene 56      gene 66      gene 69      gene 77      gene 88
#> gene 4   1.421229e-06 1.520465e-27 3.262925e-27 1.878551e-07 1.000000e+00
#> gene 5   1.000000e+00 3.618500e-01 3.889576e-01 1.000000e+00 1.000000e+00
#> gene 7   1.406181e-48 3.871476e-55 3.814918e-55 1.020476e-48 1.118920e-12
#> gene 29  2.661362e-56 9.537272e-57 9.933801e-57 2.142799e-56 5.185189e-55
#> gene 30  4.503478e-10 2.860128e-32 4.642122e-32 4.326071e-11 1.000000e+00
#> gene 31  1.000000e+00 3.155851e-01 3.458599e-01 1.000000e+00 1.000000e+00
#> gene 32  3.446275e-06 1.806902e-26 3.892678e-26 5.985653e-07 1.000000e+00
#> gene 33  5.489839e-41 1.305529e-52 1.366813e-52 1.519336e-41 8.200761e-05
#> gene 34  2.201635e-05 1.703306e-26 4.385265e-26 3.348595e-06 1.000000e+00
#> gene 36  1.000000e+00 5.165404e-02 6.041283e-02 1.000000e+00 1.000000e+00
#> gene 40  1.000000e+00 6.703902e-01 6.931495e-01 1.000000e+00 1.000000e+00
#> gene 44  1.000000e+00 6.063972e-01 6.250446e-01 1.000000e+00 1.000000e+00
#> gene 47  1.000000e+00 2.088079e-01 2.411136e-01 1.000000e+00 1.000000e+00
#> gene 49  1.000000e+00 6.103907e-01 6.401707e-01 1.000000e+00 1.000000e+00
#> gene 50  1.000000e+00 1.009720e-01 1.174950e-01 1.000000e+00 1.000000e+00
#> gene 51  1.000000e+00 1.924512e-01 2.233663e-01 1.000000e+00 1.000000e+00
#> gene 52  1.000000e+00 8.727485e-02 1.048907e-01 1.000000e+00 1.000000e+00
#> gene 53  1.000000e+00 2.565635e-02 3.440507e-02 1.000000e+00 1.000000e+00
#> gene 54  1.000000e+00 3.186043e-01 3.423848e-01 1.000000e+00 1.000000e+00
#> gene 55  1.845435e-25 3.309532e-45 4.172481e-45 1.860213e-26 9.994253e-01
#> gene 56  1.000000e+00 1.188597e-13 2.540051e-13 3.500525e-01 1.000000e+00
#> gene 66  1.000000e+00 1.000000e+00 5.302130e-01 1.000000e+00 1.000000e+00
#> gene 69  1.000000e+00 4.702734e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 77  6.503864e-01 2.157029e-13 5.717675e-13 1.000000e+00 1.000000e+00
#> gene 88  1.322705e-33 1.620871e-49 1.768708e-49 2.165329e-34 1.000000e+00
#> gene 89  1.000000e+00 5.618497e-01 5.890778e-01 1.000000e+00 1.000000e+00
#> gene 91  8.235703e-20 3.106532e-42 4.495357e-42 3.737460e-21 9.999972e-01
#> gene 111 1.849720e-09 3.686504e-32 6.118775e-32 1.343979e-10 1.000000e+00
#> XOR_EZH2 4.242942e-52 5.212484e-56 5.071474e-56 3.988096e-52 3.801521e-23
#> OR_CSF3R 1.301805e-43 1.058573e-52 1.118007e-52 7.398044e-44 2.625743e-07
#>               gene 89      gene 91     gene 111     XOR_EZH2     OR_CSF3R
#> gene 4   2.414639e-28 9.999992e-01 8.749333e-01 1.000000e+00 1.000000e+00
#> gene 5   3.036297e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 7   1.762009e-55 3.211585e-24 1.024328e-36 9.999623e-01 1.586896e-02
#> gene 29  8.806722e-57 1.426321e-55 3.972375e-56 1.924352e-44 5.477097e-53
#> gene 30  3.880165e-33 9.997861e-01 3.576821e-01 1.000000e+00 1.000000e+00
#> gene 31  2.613784e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 32  2.478433e-27 9.999993e-01 8.858138e-01 1.000000e+00 1.000000e+00
#> gene 33  3.963727e-53 4.546772e-14 7.218800e-26 1.000000e+00 8.605589e-01
#> gene 34  2.181867e-27 1.000000e+00 9.707225e-01 1.000000e+00 1.000000e+00
#> gene 36  3.655774e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 40  6.152250e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 44  5.504679e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 47  1.595828e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 49  5.488945e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 50  7.498308e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 51  1.486077e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 52  6.335123e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 53  1.641445e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 54  2.656291e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 55  4.507238e-46 4.659239e-02 9.416075e-09 1.000000e+00 1.000000e+00
#> gene 56  2.808114e-14 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 66  4.386334e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 69  4.113983e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 77  4.693334e-14 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 88  3.090685e-50 2.860947e-06 1.636512e-16 1.000000e+00 9.999997e-01
#> gene 89  1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 91  4.120282e-43 1.000000e+00 5.490475e-05 1.000000e+00 1.000000e+00
#> gene 111 5.417013e-33 9.999454e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> XOR_EZH2 1.000000e+00 1.828484e-33 8.074223e-44 1.000000e+00 1.000000e+00
#> OR_CSF3R 3.892881e-53 8.311018e-18 1.000000e+00 1.000000e+00 1.000000e+00
#> 
as.confidence(test_model, conf=c('tp', 'hg'))
#> $hg
#>              gene 4 gene 5     gene 7    gene 29    gene 30 gene 31    gene 32
#> gene 4   1.00000000  0.075 0.44777328 0.73684211 0.21356275   0.075 0.21356275
#> gene 5   0.07500000  1.000 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 7   0.44777328  0.175 1.00000000 0.96471704 0.44777328   0.000 0.07368421
#> gene 29  0.73684211  0.350 0.96471704 1.00000000 0.03684211   0.350 0.73684211
#> gene 30  0.21356275  0.075 0.44777328 0.03684211 1.00000000   0.075 0.21356275
#> gene 31  0.07500000  0.025 0.00000000 0.35000000 0.07500000   1.000 0.07500000
#> gene 32  0.21356275  0.075 0.07368421 0.73684211 0.21356275   0.075 1.00000000
#> gene 33  0.05364372  0.150 0.27862275 0.94001897 0.05364372   0.150 0.39433198
#> gene 34  0.21356275  0.075 0.44777328 0.27631579 0.21356275   0.075 0.21356275
#> gene 36  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 40  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 44  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 47  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 49  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 50  0.07500000  0.025 0.17500000 0.00000000 0.07500000   0.025 0.07500000
#> gene 51  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 52  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 53  0.07500000  0.025 0.17500000 0.35000000 0.00000000   0.025 0.07500000
#> gene 54  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 55  0.27732794  0.100 0.55224860 0.11450925 0.27732794   0.100 0.27732794
#> gene 56  0.00000000  0.050 0.32307692 0.58333333 0.14615385   0.050 0.14615385
#> gene 66  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 69  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 77  0.14615385  0.050 0.32307692 0.58333333 0.14615385   0.050 0.14615385
#> gene 88  0.33755061  0.125 0.63931138 0.22238028 0.33755061   0.125 0.33755061
#> gene 89  0.07500000  0.025 0.17500000 0.35000000 0.07500000   0.025 0.07500000
#> gene 91  0.27732794  0.100 0.55224860 0.00000000 0.27732794   0.100 0.27732794
#> gene 111 0.21356275  0.075 0.44777328 0.03684211 0.21356275   0.075 0.21356275
#> XOR_EZH2 0.09635628  1.000 0.12817509 0.97968557 0.09635628   0.200 0.49797571
#> OR_CSF3R 0.39433198  0.150 0.71144910 0.09909910 0.39433198   0.150 0.39433198
#>             gene 33    gene 34 gene 36 gene 40 gene 44 gene 47 gene 49 gene 50
#> gene 4   0.05364372 0.21356275   0.075   0.075   0.075   0.075   0.075   0.075
#> gene 5   0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 7   0.27862275 0.44777328   0.175   0.175   0.175   0.175   0.175   0.175
#> gene 29  0.94001897 0.27631579   0.350   0.350   0.350   0.350   0.350   0.000
#> gene 30  0.05364372 0.21356275   0.075   0.075   0.075   0.075   0.075   0.075
#> gene 31  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 32  0.39433198 0.21356275   0.075   0.075   0.075   0.075   0.075   0.075
#> gene 33  1.00000000 0.39433198   0.150   0.150   0.150   0.150   0.150   0.150
#> gene 34  0.39433198 1.00000000   0.075   0.075   0.075   0.075   0.075   0.075
#> gene 36  0.15000000 0.07500000   1.000   0.025   0.025   0.025   0.025   0.025
#> gene 40  0.15000000 0.07500000   0.025   1.000   0.025   0.025   0.025   0.025
#> gene 44  0.15000000 0.07500000   0.025   0.025   1.000   0.025   0.025   0.025
#> gene 47  0.15000000 0.07500000   0.025   0.025   0.025   1.000   0.025   0.025
#> gene 49  0.15000000 0.07500000   0.025   0.025   0.025   0.025   1.000   0.025
#> gene 50  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   1.000
#> gene 51  0.00000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 52  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 53  0.00000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 54  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 55  0.49254842 0.27732794   0.100   0.100   0.100   0.100   0.100   0.100
#> gene 56  0.01923077 0.14615385   0.050   0.050   0.050   0.050   0.050   0.050
#> gene 66  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 69  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 77  0.28076923 0.14615385   0.050   0.050   0.050   0.000   0.050   0.050
#> gene 88  0.57712368 0.33755061   0.125   0.125   0.125   0.125   0.125   0.125
#> gene 89  0.15000000 0.07500000   0.025   0.025   0.025   0.025   0.025   0.025
#> gene 91  0.49254842 0.02226721   0.100   0.100   0.100   0.100   0.100   0.000
#> gene 111 0.39433198 0.21356275   0.075   0.075   0.075   0.075   0.075   0.075
#> XOR_EZH2 1.00000000 0.49797571   0.200   0.200   0.200   0.200   0.200   0.200
#> OR_CSF3R 0.64961677 0.39433198   0.150   0.150   0.150   0.150   0.150   0.150
#>          gene 51 gene 52 gene 53 gene 54    gene 55    gene 56 gene 66 gene 69
#> gene 4     0.075   0.075   0.075   0.075 0.27732794 0.00000000   0.075   0.075
#> gene 5     0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 7     0.175   0.175   0.175   0.175 0.55224860 0.32307692   0.175   0.175
#> gene 29    0.350   0.350   0.350   0.350 0.11450925 0.58333333   0.350   0.350
#> gene 30    0.075   0.075   0.000   0.075 0.27732794 0.14615385   0.075   0.075
#> gene 31    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 32    0.075   0.075   0.075   0.075 0.27732794 0.14615385   0.075   0.075
#> gene 33    0.000   0.150   0.000   0.150 0.49254842 0.01923077   0.150   0.150
#> gene 34    0.075   0.075   0.075   0.075 0.27732794 0.14615385   0.075   0.075
#> gene 36    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 40    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 44    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 47    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 49    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 50    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 51    1.000   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 52    0.025   1.000   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 53    0.025   0.025   1.000   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 54    0.025   0.025   0.025   1.000 0.10000000 0.05000000   0.025   0.025
#> gene 55    0.100   0.100   0.100   0.100 1.00000000 0.19230769   0.100   0.100
#> gene 56    0.050   0.050   0.050   0.050 0.19230769 1.00000000   0.050   0.050
#> gene 66    0.025   0.025   0.025   0.025 0.10000000 0.05000000   1.000   0.025
#> gene 69    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   1.000
#> gene 77    0.050   0.050   0.050   0.050 0.19230769 0.09871795   0.050   0.050
#> gene 88    0.125   0.125   0.125   0.125 0.06899004 0.23717949   0.125   0.125
#> gene 89    0.025   0.025   0.025   0.025 0.10000000 0.05000000   0.025   0.025
#> gene 91    0.100   0.100   0.100   0.100 0.35545464 0.19230769   0.100   0.100
#> gene 111   0.075   0.075   0.075   0.075 0.02226721 0.14615385   0.075   0.075
#> XOR_EZH2   0.000   0.200   0.000   0.200 0.60652150 0.03589744   0.200   0.200
#> OR_CSF3R   0.150   0.150   0.150   0.150 1.00000000 0.28076923   0.150   0.150
#>             gene 77    gene 88 gene 89    gene 91   gene 111   XOR_EZH2
#> gene 4   0.14615385 0.33755061   0.075 0.27732794 0.21356275 0.09635628
#> gene 5   0.05000000 0.12500000   0.025 0.10000000 0.07500000 1.00000000
#> gene 7   0.32307692 0.63931138   0.175 0.55224860 0.44777328 0.12817509
#> gene 29  0.58333333 0.22238028   0.350 0.00000000 0.03684211 0.97968557
#> gene 30  0.14615385 0.33755061   0.075 0.27732794 0.21356275 0.09635628
#> gene 31  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 32  0.14615385 0.33755061   0.075 0.27732794 0.21356275 0.49797571
#> gene 33  0.28076923 0.57712368   0.150 0.49254842 0.39433198 1.00000000
#> gene 34  0.14615385 0.33755061   0.075 0.02226721 0.21356275 0.49797571
#> gene 36  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 40  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 44  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 47  0.00000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 49  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 50  0.05000000 0.12500000   0.025 0.00000000 0.07500000 0.20000000
#> gene 51  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.00000000
#> gene 52  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 53  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.00000000
#> gene 54  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 55  0.19230769 0.06899004   0.100 0.35545464 0.02226721 0.60652150
#> gene 56  0.09871795 0.23717949   0.050 0.19230769 0.14615385 0.03589744
#> gene 66  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 69  0.05000000 0.12500000   0.025 0.10000000 0.07500000 0.20000000
#> gene 77  1.00000000 0.23717949   0.050 0.19230769 0.14615385 0.36410256
#> gene 88  0.23717949 1.00000000   0.125 0.42707080 0.33755061 0.69396117
#> gene 89  0.05000000 0.12500000   1.000 0.10000000 0.07500000 1.00000000
#> gene 91  0.19230769 0.42707080   0.100 1.00000000 0.27732794 0.60652150
#> gene 111 0.14615385 0.33755061   0.075 0.27732794 1.00000000 0.49797571
#> XOR_EZH2 0.36410256 0.69396117   1.000 0.60652150 0.49797571 1.00000000
#> OR_CSF3R 0.28076923 0.15424736   0.150 0.49254842 1.00000000 1.00000000
#>           OR_CSF3R
#> gene 4   0.3943320
#> gene 5   0.1500000
#> gene 7   0.7114491
#> gene 29  0.0990991
#> gene 30  0.3943320
#> gene 31  0.1500000
#> gene 32  0.3943320
#> gene 33  0.6496168
#> gene 34  0.3943320
#> gene 36  0.1500000
#> gene 40  0.1500000
#> gene 44  0.1500000
#> gene 47  0.1500000
#> gene 49  0.1500000
#> gene 50  0.1500000
#> gene 51  0.1500000
#> gene 52  0.1500000
#> gene 53  0.1500000
#> gene 54  0.1500000
#> gene 55  1.0000000
#> gene 56  0.2807692
#> gene 66  0.1500000
#> gene 69  0.1500000
#> gene 77  0.2807692
#> gene 88  0.1542474
#> gene 89  0.1500000
#> gene 91  0.4925484
#> gene 111 1.0000000
#> XOR_EZH2 1.0000000
#> OR_CSF3R 1.0000000
#> 
#> $tp
#>                gene 4       gene 5       gene 7 gene 29      gene 30
#> gene 4   1.000000e+00 1.455680e-26 1.000000e+00       1 9.251274e-01
#> gene 5   1.000000e+00 1.000000e+00 1.000000e+00       1 1.000000e+00
#> gene 7   5.329851e-38 3.926171e-55 1.000000e+00       1 1.724654e-36
#> gene 29  4.736819e-56 9.709084e-57 6.798626e-51       1 4.357290e-56
#> gene 30  7.503856e-02 2.295716e-31 1.000000e+00       1 1.000000e+00
#> gene 31  1.000000e+00 4.509268e-01 1.000000e+00       1 1.000000e+00
#> gene 32  5.489782e-01 1.282066e-25 1.000000e+00       1 9.344646e-01
#> gene 33  9.373014e-28 1.918776e-52 9.991563e-01       1 3.120373e-25
#> gene 34  7.643872e-01 1.464270e-25 1.000000e+00       1 9.860117e-01
#> gene 36  1.000000e+00 1.002462e-01 1.000000e+00       1 1.000000e+00
#> gene 40  1.000000e+00 7.909232e-01 1.000000e+00       1 1.000000e+00
#> gene 44  1.000000e+00 7.337264e-01 1.000000e+00       1 1.000000e+00
#> gene 47  1.000000e+00 3.174193e-01 1.000000e+00       1 1.000000e+00
#> gene 49  1.000000e+00 7.387701e-01 1.000000e+00       1 1.000000e+00
#> gene 50  1.000000e+00 1.779009e-01 1.000000e+00       1 1.000000e+00
#> gene 51  1.000000e+00 3.036831e-01 1.000000e+00       1 1.000000e+00
#> gene 52  1.000000e+00 1.593646e-01 1.000000e+00       1 1.000000e+00
#> gene 53  1.000000e+00 5.479794e-02 1.000000e+00       1 1.000000e+00
#> gene 54  1.000000e+00 4.507570e-01 1.000000e+00       1 1.000000e+00
#> gene 55  8.972807e-11 8.857734e-45 1.000000e+00       1 5.160503e-08
#> gene 56  9.999986e-01 8.703374e-13 1.000000e+00       1 1.000000e+00
#> gene 66  1.000000e+00 6.386084e-01 1.000000e+00       1 1.000000e+00
#> gene 69  1.000000e+00 6.115105e-01 1.000000e+00       1 1.000000e+00
#> gene 77  9.999998e-01 1.552462e-12 1.000000e+00       1 1.000000e+00
#> gene 88  1.126419e-18 2.950660e-49 1.000000e+00       1 9.388325e-16
#> gene 89  1.000000e+00 6.967985e-01 1.000000e+00       1 1.000000e+00
#> gene 91  8.337071e-07 1.622532e-41 1.000000e+00       1 2.148264e-04
#> gene 111 1.253088e-01 3.862890e-31 1.000000e+00       1 6.427562e-01
#> XOR_EZH2 1.657722e-44 1.000000e+00 3.787203e-05       1 1.100280e-43
#> OR_CSF3R 9.610599e-32 1.261960e-52 9.841773e-01       1 6.816777e-30
#>               gene 31      gene 32      gene 33      gene 34      gene 36
#> gene 4   1.547314e-26 4.514859e-01 1.000000e+00 2.359743e-01 2.342545e-22
#> gene 5   5.495566e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.999670e-01
#> gene 7   4.356112e-55 2.278343e-38 8.470797e-04 4.899826e-40 2.294724e-54
#> gene 29  1.064251e-56 3.984755e-56 6.509270e-52 4.145089e-56 1.447628e-56
#> gene 30  1.985077e-31 6.568508e-02 1.000000e+00 1.403023e-02 2.436703e-27
#> gene 31  1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 8.809486e-01
#> gene 32  1.764283e-25 1.000000e+00 1.000000e+00 2.823813e-01 1.551439e-21
#> gene 33  1.847050e-52 4.563330e-28 1.000000e+00 3.690851e-30 6.286873e-51
#> gene 34  2.002835e-25 7.180153e-01 1.000000e+00 1.000000e+00 3.799642e-21
#> gene 36  1.192927e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 40  8.227594e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.813282e-01
#> gene 44  7.662783e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.661370e-01
#> gene 47  3.798879e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.121589e-01
#> gene 49  7.806570e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.737212e-01
#> gene 50  2.095278e-01 1.000000e+00 1.000000e+00 1.000000e+00 6.505522e-01
#> gene 51  3.565901e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.052310e-01
#> gene 52  1.911901e-01 1.000000e+00 1.000000e+00 1.000000e+00 6.442144e-01
#> gene 53  7.289301e-02 1.000000e+00 1.000000e+00 1.000000e+00 4.286338e-01
#> gene 54  4.948992e-01 1.000000e+00 1.000000e+00 1.000000e+00 8.682621e-01
#> gene 55  8.353704e-45 5.618501e-11 1.000000e+00 4.545591e-13 7.177829e-42
#> gene 56  1.321968e-12 9.999966e-01 1.000000e+00 9.999781e-01 2.419613e-09
#> gene 66  6.848490e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.484744e-01
#> gene 69  6.545899e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.397323e-01
#> gene 77  2.963835e-12 9.999994e-01 1.000000e+00 9.999967e-01 6.503131e-09
#> gene 88  2.781146e-49 7.497944e-19 9.999184e-01 3.539068e-21 4.251280e-47
#> gene 89  7.390193e-01 1.000000e+00 1.000000e+00 1.000000e+00 9.635395e-01
#> gene 91  1.150112e-41 6.597718e-07 1.000000e+00 1.650994e-08 6.652005e-38
#> gene 111 2.698640e-31 1.144128e-01 1.000000e+00 2.935592e-02 5.982824e-27
#> XOR_EZH2 5.535303e-56 5.693025e-45 1.000000e+00 4.306813e-46 1.354017e-55
#> OR_CSF3R 1.483475e-52 4.947292e-32 1.396996e-01 6.632084e-34 1.866588e-51
#>               gene 40      gene 44      gene 47      gene 49      gene 50
#> gene 4   2.165386e-29 5.008267e-27 2.263522e-26 3.192407e-29 5.277778e-24
#> gene 5   2.094293e-01 2.666752e-01 6.830185e-01 2.616288e-01 8.224158e-01
#> gene 7   9.042160e-56 4.931317e-55 4.003732e-55 1.159402e-55 1.210707e-54
#> gene 29  8.419176e-57 9.175701e-57 6.888185e-57 8.336787e-57 1.287253e-56
#> gene 30  2.788036e-34 1.056706e-31 5.157664e-31 5.227785e-34 5.474974e-29
#> gene 31  1.775588e-01 2.340961e-01 6.205786e-01 2.197051e-01 7.908218e-01
#> gene 32  2.086112e-28 3.758638e-26 3.812068e-25 4.301145e-28 4.958266e-23
#> gene 33  1.214608e-53 2.709216e-52 2.057515e-52 1.677847e-53 1.443304e-51
#> gene 34  1.738248e-28 4.834981e-26 3.011753e-25 3.060859e-28 8.616538e-23
#> gene 36  1.872742e-02 3.395435e-02 1.881696e-01 2.635303e-02 3.498953e-01
#> gene 40  1.000000e+00 5.643395e-01 9.014077e-01 5.666406e-01 9.577996e-01
#> gene 44  4.361437e-01 1.000000e+00 8.638005e-01 5.024467e-01 9.314922e-01
#> gene 47  9.880598e-02 1.364671e-01 1.000000e+00 1.324596e-01 7.019801e-01
#> gene 49  4.338431e-01 4.980426e-01 8.678041e-01 1.000000e+00 9.435145e-01
#> gene 50  4.231026e-02 6.866853e-02 2.984434e-01 5.662415e-02 1.000000e+00
#> gene 51  9.189371e-02 1.329884e-01 4.624964e-01 1.189435e-01 6.829473e-01
#> gene 52  3.484817e-02 5.951470e-02 2.686340e-01 4.657274e-02 4.859700e-01
#> gene 53  7.824031e-03 1.625413e-02 1.062676e-01 1.103062e-02 2.761876e-01
#> gene 54  1.806809e-01 2.360790e-01 6.260883e-01 2.254798e-01 7.775074e-01
#> gene 55  5.230203e-47 1.126160e-44 1.553198e-44 9.837351e-47 4.491146e-43
#> gene 56  3.754344e-15 1.360701e-13 2.654492e-12 6.689968e-15 1.688733e-10
#> gene 66  3.300539e-01 3.940738e-01 7.915445e-01 3.900800e-01 8.992429e-01
#> gene 69  3.072810e-01 3.754195e-01 7.592675e-01 3.602876e-01 8.827444e-01
#> gene 77  6.488969e-15 2.216224e-13 4.511947e-12 1.134950e-14 4.576681e-10
#> gene 88  5.425064e-51 4.217177e-49 4.953952e-49 9.833821e-51 5.314708e-48
#> gene 89  3.852445e-01 4.500171e-01 8.407157e-01 4.515920e-01 9.251892e-01
#> gene 91  3.474747e-44 1.925081e-41 2.180308e-41 5.755818e-44 1.661066e-39
#> gene 111 3.779886e-34 2.032201e-31 5.878879e-31 5.866015e-34 9.465023e-29
#> XOR_EZH2 2.021003e-56 5.439267e-56 4.632102e-56 2.432208e-56 9.923625e-56
#> OR_CSF3R 1.546387e-53 1.361436e-52 1.636824e-52 2.239969e-53 7.180214e-52
#>               gene 51      gene 52      gene 53      gene 54      gene 55
#> gene 4   1.086371e-26 4.166290e-25 4.513530e-24 3.046136e-25 1.000000e+00
#> gene 5   6.967445e-01 8.409308e-01 9.453371e-01 5.497257e-01 1.000000e+00
#> gene 7   2.633366e-55 3.104739e-55 4.347804e-55 1.018560e-54 1.883964e-22
#> gene 29  9.352671e-57 1.181415e-56 1.069432e-56 1.136122e-56 6.828657e-56
#> gene 30  1.128523e-31 2.456006e-30 2.827924e-29 5.478567e-30 9.999999e-01
#> gene 31  6.438652e-01 8.091411e-01 9.272754e-01 5.055866e-01 1.000000e+00
#> gene 32  1.712602e-25 4.817431e-24 7.261810e-23 2.539887e-24 1.000000e+00
#> gene 33  8.126240e-53 1.574250e-52 3.167398e-52 1.025199e-51 1.663992e-11
#> gene 34  1.595057e-25 7.199230e-24 1.048880e-22 3.874618e-24 1.000000e+00
#> gene 36  1.951038e-01 3.562367e-01 5.718416e-01 1.319965e-01 1.000000e+00
#> gene 40  9.083082e-01 9.652457e-01 9.922022e-01 8.196405e-01 1.000000e+00
#> gene 44  8.672738e-01 9.406292e-01 9.837951e-01 7.642969e-01 1.000000e+00
#> gene 47  5.379924e-01 7.317696e-01 8.939573e-01 3.743749e-01 1.000000e+00
#> gene 49  8.813001e-01 9.535461e-01 9.890048e-01 7.748872e-01 1.000000e+00
#> gene 50  3.174865e-01 5.145136e-01 7.242182e-01 2.228543e-01 1.000000e+00
#> gene 51  1.000000e+00 7.069561e-01 8.729173e-01 3.628007e-01 1.000000e+00
#> gene 52  2.934635e-01 1.000000e+00 7.143980e-01 2.063617e-01 1.000000e+00
#> gene 53  1.273372e-01 2.860155e-01 1.000000e+00 8.201938e-02 1.000000e+00
#> gene 54  6.376563e-01 7.939845e-01 9.181644e-01 1.000000e+00 1.000000e+00
#> gene 55  2.928759e-45 1.450209e-44 7.671125e-44 1.599977e-43 1.000000e+00
#> gene 56  2.391171e-12 5.531515e-11 6.047829e-10 6.903869e-12 1.000000e+00
#> gene 66  8.078831e-01 9.129183e-01 9.744163e-01 6.818311e-01 1.000000e+00
#> gene 69  7.769980e-01 8.953302e-01 9.656875e-01 6.580630e-01 1.000000e+00
#> gene 77  5.924895e-12 1.791474e-10 2.121713e-09 1.367533e-11 1.000000e+00
#> gene 88  1.083763e-49 2.840822e-49 1.005260e-48 3.058781e-48 5.771054e-04
#> gene 89  8.516760e-01 9.368002e-01 9.836354e-01 7.347714e-01 1.000000e+00
#> gene 91  3.268978e-42 4.019945e-41 2.577904e-40 4.730861e-40 9.535215e-01
#> gene 111 1.201388e-31 3.564199e-30 3.490003e-29 1.059821e-29 1.000000e+00
#> XOR_EZH2 3.987969e-56 4.161263e-56 4.919078e-56 8.725395e-56 1.255183e-32
#> OR_CSF3R 9.498741e-53 1.490719e-52 2.977122e-52 4.358181e-52 1.000000e+00
#>               gene 56      gene 66      gene 69      gene 77      gene 88
#> gene 4   1.421229e-06 1.520465e-27 3.262925e-27 1.878551e-07 1.000000e+00
#> gene 5   1.000000e+00 3.618500e-01 3.889576e-01 1.000000e+00 1.000000e+00
#> gene 7   1.406181e-48 3.871476e-55 3.814918e-55 1.020476e-48 1.118920e-12
#> gene 29  2.661362e-56 9.537272e-57 9.933801e-57 2.142799e-56 5.185189e-55
#> gene 30  4.503478e-10 2.860128e-32 4.642122e-32 4.326071e-11 1.000000e+00
#> gene 31  1.000000e+00 3.155851e-01 3.458599e-01 1.000000e+00 1.000000e+00
#> gene 32  3.446275e-06 1.806902e-26 3.892678e-26 5.985653e-07 1.000000e+00
#> gene 33  5.489839e-41 1.305529e-52 1.366813e-52 1.519336e-41 8.200761e-05
#> gene 34  2.201635e-05 1.703306e-26 4.385265e-26 3.348595e-06 1.000000e+00
#> gene 36  1.000000e+00 5.165404e-02 6.041283e-02 1.000000e+00 1.000000e+00
#> gene 40  1.000000e+00 6.703902e-01 6.931495e-01 1.000000e+00 1.000000e+00
#> gene 44  1.000000e+00 6.063972e-01 6.250446e-01 1.000000e+00 1.000000e+00
#> gene 47  1.000000e+00 2.088079e-01 2.411136e-01 1.000000e+00 1.000000e+00
#> gene 49  1.000000e+00 6.103907e-01 6.401707e-01 1.000000e+00 1.000000e+00
#> gene 50  1.000000e+00 1.009720e-01 1.174950e-01 1.000000e+00 1.000000e+00
#> gene 51  1.000000e+00 1.924512e-01 2.233663e-01 1.000000e+00 1.000000e+00
#> gene 52  1.000000e+00 8.727485e-02 1.048907e-01 1.000000e+00 1.000000e+00
#> gene 53  1.000000e+00 2.565635e-02 3.440507e-02 1.000000e+00 1.000000e+00
#> gene 54  1.000000e+00 3.186043e-01 3.423848e-01 1.000000e+00 1.000000e+00
#> gene 55  1.845435e-25 3.309532e-45 4.172481e-45 1.860213e-26 9.994253e-01
#> gene 56  1.000000e+00 1.188597e-13 2.540051e-13 3.500525e-01 1.000000e+00
#> gene 66  1.000000e+00 1.000000e+00 5.302130e-01 1.000000e+00 1.000000e+00
#> gene 69  1.000000e+00 4.702734e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 77  6.503864e-01 2.157029e-13 5.717675e-13 1.000000e+00 1.000000e+00
#> gene 88  1.322705e-33 1.620871e-49 1.768708e-49 2.165329e-34 1.000000e+00
#> gene 89  1.000000e+00 5.618497e-01 5.890778e-01 1.000000e+00 1.000000e+00
#> gene 91  8.235703e-20 3.106532e-42 4.495357e-42 3.737460e-21 9.999972e-01
#> gene 111 1.849720e-09 3.686504e-32 6.118775e-32 1.343979e-10 1.000000e+00
#> XOR_EZH2 4.242942e-52 5.212484e-56 5.071474e-56 3.988096e-52 3.801521e-23
#> OR_CSF3R 1.301805e-43 1.058573e-52 1.118007e-52 7.398044e-44 2.625743e-07
#>               gene 89      gene 91     gene 111     XOR_EZH2     OR_CSF3R
#> gene 4   2.414639e-28 9.999992e-01 8.749333e-01 1.000000e+00 1.000000e+00
#> gene 5   3.036297e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 7   1.762009e-55 3.211585e-24 1.024328e-36 9.999623e-01 1.586896e-02
#> gene 29  8.806722e-57 1.426321e-55 3.972375e-56 1.924352e-44 5.477097e-53
#> gene 30  3.880165e-33 9.997861e-01 3.576821e-01 1.000000e+00 1.000000e+00
#> gene 31  2.613784e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 32  2.478433e-27 9.999993e-01 8.858138e-01 1.000000e+00 1.000000e+00
#> gene 33  3.963727e-53 4.546772e-14 7.218800e-26 1.000000e+00 8.605589e-01
#> gene 34  2.181867e-27 1.000000e+00 9.707225e-01 1.000000e+00 1.000000e+00
#> gene 36  3.655774e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 40  6.152250e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 44  5.504679e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 47  1.595828e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 49  5.488945e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 50  7.498308e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 51  1.486077e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 52  6.335123e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 53  1.641445e-02 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 54  2.656291e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 55  4.507238e-46 4.659239e-02 9.416075e-09 1.000000e+00 1.000000e+00
#> gene 56  2.808114e-14 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 66  4.386334e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 69  4.113983e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 77  4.693334e-14 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 88  3.090685e-50 2.860947e-06 1.636512e-16 1.000000e+00 9.999997e-01
#> gene 89  1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00
#> gene 91  4.120282e-43 1.000000e+00 5.490475e-05 1.000000e+00 1.000000e+00
#> gene 111 5.417013e-33 9.999454e-01 1.000000e+00 1.000000e+00 1.000000e+00
#> XOR_EZH2 1.000000e+00 1.828484e-33 8.074223e-44 1.000000e+00 1.000000e+00
#> OR_CSF3R 3.892881e-53 8.311018e-18 1.000000e+00 1.000000e+00 1.000000e+00
#>