as.joint.probs

Extract the joint probabilities from a TRONCO model. The return matrix is indexed with rownames/colnames which represent genotype keys - these can be resolved with function keysToNames. It is possible to specify a subset of events to build the matrix, a subset of models if multiple reconstruction have been performed. Also, either the observed or fit probabilities can be extracted.

as.joint.probs(
  x,
  events = as.events(x),
  models = names(x$model),
  type = "observed"
)

Arguments

x
A TRONCO model.

events
A subset of events as of as.events(x), all by default.

models
A subset of reconstructed models, all by default.

type
observed

Value

The joint probabilities in a TRONCO model.

Examples

data(test_model)
as.joint.probs(test_model)

#> $capri_bic
#>              gene 4     gene 5     gene 7    gene 29    gene 30 gene 31
#> gene 4   0.07348485 0.00000000 0.02272727 0.00000000 0.00000000   0.000
#> gene 5   0.00000000 0.02484848 0.02484848 0.00000000 0.00000000   0.000
#> gene 7   0.02272727 0.02484848 0.16984848 0.02439394 0.00000000   0.025
#> gene 29  0.00000000 0.00000000 0.02439394 0.35257576 0.05090909   0.000
#> gene 30  0.00000000 0.00000000 0.00000000 0.05090909 0.07909091   0.000
#> gene 31  0.00000000 0.00000000 0.02500000 0.00000000 0.00000000   0.025
#> gene 32  0.00000000 0.02484848 0.04863636 0.00000000 0.00000000   0.000
#> gene 33  0.04939394 0.00000000 0.04696970 0.00000000 0.02818182   0.000
#> gene 34  0.02272727 0.00000000 0.02272727 0.04727273 0.00000000   0.000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 44  0.00000000 0.00000000 0.00000000 0.02378788 0.02378788   0.000
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 49  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 50  0.00000000 0.00000000 0.00000000 0.02742424 0.00000000   0.000
#> gene 51  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.02818182   0.000
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 55  0.00000000 0.00000000 0.00000000 0.05590909 0.00000000   0.000
#> gene 56  0.05075758 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 66  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 69  0.02409091 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 77  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 88  0.00000000 0.00000000 0.02424242 0.05075758 0.02378788   0.000
#> gene 89  0.00000000 0.00000000 0.00000000 0.02318182 0.00000000   0.000
#> gene 91  0.00000000 0.00000000 0.00000000 0.10000000 0.00000000   0.000
#> gene 111 0.00000000 0.00000000 0.00000000 0.05287879 0.00000000   0.000
#> XOR_EZH2 0.04939394 0.02484848 0.07181818 0.02318182 0.02818182   0.000
#> OR_CSF3R 0.00000000 0.00000000 0.00000000 0.07984848 0.00000000   0.000
#>             gene 32    gene 33    gene 34    gene 36    gene 40    gene 44
#> gene 4   0.00000000 0.04939394 0.02272727 0.00000000 0.00000000 0.00000000
#> gene 5   0.02484848 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 7   0.04863636 0.04696970 0.02272727 0.00000000 0.00000000 0.00000000
#> gene 29  0.00000000 0.00000000 0.04727273 0.00000000 0.00000000 0.02378788
#> gene 30  0.00000000 0.02818182 0.00000000 0.00000000 0.00000000 0.02378788
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.07257576 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 33  0.00000000 0.14984848 0.02272727 0.00000000 0.00000000 0.00000000
#> gene 34  0.00000000 0.02272727 0.07000000 0.00000000 0.00000000 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.02878788 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.02212121 0.00000000
#> gene 44  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02378788
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49  0.00000000 0.02257576 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 50  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.02545455 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 53  0.00000000 0.02818182 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 56  0.00000000 0.02666667 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 66  0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 69  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 77  0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 88  0.00000000 0.02424242 0.00000000 0.00000000 0.00000000 0.02378788
#> gene 89  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 91  0.00000000 0.00000000 0.04727273 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> XOR_EZH2 0.02484848 0.14984848 0.02272727 0.00000000 0.00000000 0.00000000
#> OR_CSF3R 0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             gene 47    gene 49    gene 50    gene 51    gene 52    gene 53
#> gene 4   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 5   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 7   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 29  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 30  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02818182
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 33  0.00000000 0.02257576 0.00000000 0.02545455 0.00000000 0.02818182
#> gene 34  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47  0.02590909 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49  0.00000000 0.02257576 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 50  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.00000000 0.00000000 0.02545455 0.00000000 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.02666667 0.00000000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02818182
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 56  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 66  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 69  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 77  0.02590909 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 88  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 89  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 91  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> XOR_EZH2 0.00000000 0.02257576 0.00000000 0.02545455 0.00000000 0.02818182
#> OR_CSF3R 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             gene 54    gene 55    gene 56    gene 66    gene 69    gene 77
#> gene 4   0.00000000 0.00000000 0.05075758 0.00000000 0.02409091 0.00000000
#> gene 5   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 7   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 29  0.00000000 0.05590909 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 30  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.00000000 0.02393939 0.00000000 0.02393939 0.00000000 0.02393939
#> gene 33  0.00000000 0.00000000 0.02666667 0.00000000 0.00000000 0.00000000
#> gene 34  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02590909
#> gene 49  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 50  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 54  0.02575758 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.00000000 0.10590909 0.00000000 0.02393939 0.00000000 0.02393939
#> gene 56  0.00000000 0.00000000 0.05075758 0.00000000 0.02409091 0.00000000
#> gene 66  0.00000000 0.02393939 0.00000000 0.02393939 0.00000000 0.02393939
#> gene 69  0.00000000 0.00000000 0.02409091 0.00000000 0.02409091 0.00000000
#> gene 77  0.00000000 0.02393939 0.00000000 0.02393939 0.00000000 0.04984848
#> gene 88  0.00000000 0.02696970 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 89  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 91  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.02893939 0.00000000 0.00000000 0.00000000 0.00000000
#> XOR_EZH2 0.00000000 0.00000000 0.02666667 0.00000000 0.00000000 0.00000000
#> OR_CSF3R 0.00000000 0.10590909 0.00000000 0.02393939 0.00000000 0.02393939
#>             gene 88    gene 89    gene 91   gene 111   XOR_EZH2   OR_CSF3R
#> gene 4   0.00000000 0.00000000 0.00000000 0.00000000 0.04939394 0.00000000
#> gene 5   0.00000000 0.00000000 0.00000000 0.00000000 0.02484848 0.00000000
#> gene 7   0.02424242 0.00000000 0.00000000 0.00000000 0.07181818 0.00000000
#> gene 29  0.05075758 0.02318182 0.10000000 0.05287879 0.02318182 0.07984848
#> gene 30  0.02378788 0.00000000 0.00000000 0.00000000 0.02818182 0.00000000
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.00000000 0.00000000 0.00000000 0.00000000 0.02484848 0.02393939
#> gene 33  0.02424242 0.00000000 0.00000000 0.00000000 0.14984848 0.00000000
#> gene 34  0.00000000 0.00000000 0.04727273 0.00000000 0.02272727 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44  0.02378788 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49  0.00000000 0.00000000 0.00000000 0.00000000 0.02257576 0.00000000
#> gene 50  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.00000000 0.00000000 0.00000000 0.02545455 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.02818182 0.00000000
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.02696970 0.00000000 0.00000000 0.02893939 0.00000000 0.10590909
#> gene 56  0.00000000 0.00000000 0.00000000 0.00000000 0.02666667 0.00000000
#> gene 66  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02393939
#> gene 69  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 77  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02393939
#> gene 88  0.12439394 0.00000000 0.00000000 0.00000000 0.02424242 0.02696970
#> gene 89  0.00000000 0.02318182 0.00000000 0.00000000 0.02318182 0.00000000
#> gene 91  0.00000000 0.00000000 0.10000000 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.00000000 0.00000000 0.07803030 0.00000000 0.07803030
#> XOR_EZH2 0.02424242 0.02318182 0.00000000 0.00000000 0.19787879 0.00000000
#> OR_CSF3R 0.02696970 0.00000000 0.00000000 0.07803030 0.00000000 0.15500000
#> 
#> $capri_aic
#>              gene 4     gene 5     gene 7    gene 29    gene 30 gene 31
#> gene 4   0.07348485 0.00000000 0.02272727 0.00000000 0.00000000   0.000
#> gene 5   0.00000000 0.02484848 0.02484848 0.00000000 0.00000000   0.000
#> gene 7   0.02272727 0.02484848 0.16984848 0.02439394 0.00000000   0.025
#> gene 29  0.00000000 0.00000000 0.02439394 0.35257576 0.05090909   0.000
#> gene 30  0.00000000 0.00000000 0.00000000 0.05090909 0.07909091   0.000
#> gene 31  0.00000000 0.00000000 0.02500000 0.00000000 0.00000000   0.025
#> gene 32  0.00000000 0.02484848 0.04863636 0.00000000 0.00000000   0.000
#> gene 33  0.04939394 0.00000000 0.04696970 0.00000000 0.02818182   0.000
#> gene 34  0.02272727 0.00000000 0.02272727 0.04727273 0.00000000   0.000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 44  0.00000000 0.00000000 0.00000000 0.02378788 0.02378788   0.000
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 49  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 50  0.00000000 0.00000000 0.00000000 0.02742424 0.00000000   0.000
#> gene 51  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.02818182   0.000
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 55  0.00000000 0.00000000 0.00000000 0.05590909 0.00000000   0.000
#> gene 56  0.05075758 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 66  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 69  0.02409091 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 77  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000   0.000
#> gene 88  0.00000000 0.00000000 0.02424242 0.05075758 0.02378788   0.000
#> gene 89  0.00000000 0.00000000 0.00000000 0.02318182 0.00000000   0.000
#> gene 91  0.00000000 0.00000000 0.00000000 0.10000000 0.00000000   0.000
#> gene 111 0.00000000 0.00000000 0.00000000 0.05287879 0.00000000   0.000
#> XOR_EZH2 0.04939394 0.02484848 0.07181818 0.02318182 0.02818182   0.000
#> OR_CSF3R 0.00000000 0.00000000 0.00000000 0.07984848 0.00000000   0.000
#>             gene 32    gene 33    gene 34    gene 36    gene 40    gene 44
#> gene 4   0.00000000 0.04939394 0.02272727 0.00000000 0.00000000 0.00000000
#> gene 5   0.02484848 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 7   0.04863636 0.04696970 0.02272727 0.00000000 0.00000000 0.00000000
#> gene 29  0.00000000 0.00000000 0.04727273 0.00000000 0.00000000 0.02378788
#> gene 30  0.00000000 0.02818182 0.00000000 0.00000000 0.00000000 0.02378788
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.07257576 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 33  0.00000000 0.14984848 0.02272727 0.00000000 0.00000000 0.00000000
#> gene 34  0.00000000 0.02272727 0.07000000 0.00000000 0.00000000 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.02878788 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.02212121 0.00000000
#> gene 44  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02378788
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49  0.00000000 0.02257576 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 50  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.02545455 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 53  0.00000000 0.02818182 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 56  0.00000000 0.02666667 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 66  0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 69  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 77  0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 88  0.00000000 0.02424242 0.00000000 0.00000000 0.00000000 0.02378788
#> gene 89  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 91  0.00000000 0.00000000 0.04727273 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> XOR_EZH2 0.02484848 0.14984848 0.02272727 0.00000000 0.00000000 0.00000000
#> OR_CSF3R 0.02393939 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             gene 47    gene 49    gene 50    gene 51    gene 52    gene 53
#> gene 4   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 5   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 7   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 29  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 30  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02818182
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 33  0.00000000 0.02257576 0.00000000 0.02545455 0.00000000 0.02818182
#> gene 34  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47  0.02590909 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49  0.00000000 0.02257576 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 50  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.00000000 0.00000000 0.02545455 0.00000000 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.02666667 0.00000000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02818182
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 56  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 66  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 69  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 77  0.02590909 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 88  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 89  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 91  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> XOR_EZH2 0.00000000 0.02257576 0.00000000 0.02545455 0.00000000 0.02818182
#> OR_CSF3R 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             gene 54    gene 55    gene 56    gene 66    gene 69    gene 77
#> gene 4   0.00000000 0.00000000 0.05075758 0.00000000 0.02409091 0.00000000
#> gene 5   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 7   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 29  0.00000000 0.05590909 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 30  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.00000000 0.02393939 0.00000000 0.02393939 0.00000000 0.02393939
#> gene 33  0.00000000 0.00000000 0.02666667 0.00000000 0.00000000 0.00000000
#> gene 34  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02590909
#> gene 49  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 50  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 54  0.02575758 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.00000000 0.10590909 0.00000000 0.02393939 0.00000000 0.02393939
#> gene 56  0.00000000 0.00000000 0.05075758 0.00000000 0.02409091 0.00000000
#> gene 66  0.00000000 0.02393939 0.00000000 0.02393939 0.00000000 0.02393939
#> gene 69  0.00000000 0.00000000 0.02409091 0.00000000 0.02409091 0.00000000
#> gene 77  0.00000000 0.02393939 0.00000000 0.02393939 0.00000000 0.04984848
#> gene 88  0.00000000 0.02696970 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 89  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 91  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.02893939 0.00000000 0.00000000 0.00000000 0.00000000
#> XOR_EZH2 0.00000000 0.00000000 0.02666667 0.00000000 0.00000000 0.00000000
#> OR_CSF3R 0.00000000 0.10590909 0.00000000 0.02393939 0.00000000 0.02393939
#>             gene 88    gene 89    gene 91   gene 111   XOR_EZH2   OR_CSF3R
#> gene 4   0.00000000 0.00000000 0.00000000 0.00000000 0.04939394 0.00000000
#> gene 5   0.00000000 0.00000000 0.00000000 0.00000000 0.02484848 0.00000000
#> gene 7   0.02424242 0.00000000 0.00000000 0.00000000 0.07181818 0.00000000
#> gene 29  0.05075758 0.02318182 0.10000000 0.05287879 0.02318182 0.07984848
#> gene 30  0.02378788 0.00000000 0.00000000 0.00000000 0.02818182 0.00000000
#> gene 31  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32  0.00000000 0.00000000 0.00000000 0.00000000 0.02484848 0.02393939
#> gene 33  0.02424242 0.00000000 0.00000000 0.00000000 0.14984848 0.00000000
#> gene 34  0.00000000 0.00000000 0.04727273 0.00000000 0.02272727 0.00000000
#> gene 36  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44  0.02378788 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49  0.00000000 0.00000000 0.00000000 0.00000000 0.02257576 0.00000000
#> gene 50  0.00000000 0.00000000 0.02742424 0.00000000 0.00000000 0.00000000
#> gene 51  0.00000000 0.00000000 0.00000000 0.00000000 0.02545455 0.00000000
#> gene 52  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 53  0.00000000 0.00000000 0.00000000 0.00000000 0.02818182 0.00000000
#> gene 54  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 55  0.02696970 0.00000000 0.00000000 0.02893939 0.00000000 0.10590909
#> gene 56  0.00000000 0.00000000 0.00000000 0.00000000 0.02666667 0.00000000
#> gene 66  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02393939
#> gene 69  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 77  0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02393939
#> gene 88  0.12439394 0.00000000 0.00000000 0.00000000 0.02424242 0.02696970
#> gene 89  0.00000000 0.02318182 0.00000000 0.00000000 0.02318182 0.00000000
#> gene 91  0.00000000 0.00000000 0.10000000 0.00000000 0.00000000 0.00000000
#> gene 111 0.00000000 0.00000000 0.00000000 0.07803030 0.00000000 0.07803030
#> XOR_EZH2 0.02424242 0.02318182 0.00000000 0.00000000 0.19787879 0.00000000
#> OR_CSF3R 0.02696970 0.00000000 0.00000000 0.07803030 0.00000000 0.15500000
#> 

as.joint.probs(test_model, events=as.events(test_model)[5:15,])

#> $capri_bic
#>            gene 30 gene 31    gene 32    gene 33    gene 34    gene 36
#> gene 30 0.07909091   0.000 0.00000000 0.02818182 0.00000000 0.00000000
#> gene 31 0.00000000   0.025 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32 0.00000000   0.000 0.07257576 0.00000000 0.00000000 0.00000000
#> gene 33 0.02818182   0.000 0.00000000 0.14984848 0.02272727 0.00000000
#> gene 34 0.00000000   0.000 0.00000000 0.02272727 0.07000000 0.00000000
#> gene 36 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.02878788
#> gene 40 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44 0.02378788   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49 0.00000000   0.000 0.00000000 0.02257576 0.00000000 0.00000000
#> gene 50 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#>            gene 40    gene 44    gene 47    gene 49    gene 50
#> gene 30 0.00000000 0.02378788 0.00000000 0.00000000 0.00000000
#> gene 31 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 33 0.00000000 0.00000000 0.00000000 0.02257576 0.00000000
#> gene 34 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 36 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40 0.02212121 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44 0.00000000 0.02378788 0.00000000 0.00000000 0.00000000
#> gene 47 0.00000000 0.00000000 0.02590909 0.00000000 0.00000000
#> gene 49 0.00000000 0.00000000 0.00000000 0.02257576 0.00000000
#> gene 50 0.00000000 0.00000000 0.00000000 0.00000000 0.02742424
#> 
#> $capri_aic
#>            gene 30 gene 31    gene 32    gene 33    gene 34    gene 36
#> gene 30 0.07909091   0.000 0.00000000 0.02818182 0.00000000 0.00000000
#> gene 31 0.00000000   0.025 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32 0.00000000   0.000 0.07257576 0.00000000 0.00000000 0.00000000
#> gene 33 0.02818182   0.000 0.00000000 0.14984848 0.02272727 0.00000000
#> gene 34 0.00000000   0.000 0.00000000 0.02272727 0.07000000 0.00000000
#> gene 36 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.02878788
#> gene 40 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44 0.02378788   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 47 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 49 0.00000000   0.000 0.00000000 0.02257576 0.00000000 0.00000000
#> gene 50 0.00000000   0.000 0.00000000 0.00000000 0.00000000 0.00000000
#>            gene 40    gene 44    gene 47    gene 49    gene 50
#> gene 30 0.00000000 0.02378788 0.00000000 0.00000000 0.00000000
#> gene 31 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 32 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 33 0.00000000 0.00000000 0.00000000 0.02257576 0.00000000
#> gene 34 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 36 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 40 0.02212121 0.00000000 0.00000000 0.00000000 0.00000000
#> gene 44 0.00000000 0.02378788 0.00000000 0.00000000 0.00000000
#> gene 47 0.00000000 0.00000000 0.02590909 0.00000000 0.00000000
#> gene 49 0.00000000 0.00000000 0.00000000 0.02257576 0.00000000
#> gene 50 0.00000000 0.00000000 0.00000000 0.00000000 0.02742424
#>