as.kfold.prederr

Returns a dataframe with all the prediction error score in a TRONCO model. It is possible to specify a subset of events or models if multiple reconstruction have been performed.

as.kfold.prederr(
  x,
  events = as.events(x),
  models = names(x$model),
  values = FALSE,
  table = FALSE
)

Arguments

x
A TRONCO model.

events
A subset of events as of as.events(x), all by default.

models
A subset of reconstructed models, all by default.

values
If you want to see also the values

table
Keep the original table (defaul false)

Value

All the bootstrap scores in a TRONCO model

Examples

data(test_model_kfold)
as.kfold.prederr(test_model_kfold)

#> $capri_bic
#>                                   SELECTED MEAN.PREDERR  SD.PREDERR
#> 1                             ins_del TET2       0.0875 0.021245915
#> 2                             ins_del EZH2       0.0250 0.000000000
#> 3                            ins_del ASXL1       0.1750 0.000000000
#> 4          missense_point_mutations SETBP1       0.3500 0.000000000
#> 5      missense_point_mutations NRAS_Ex2_3       0.0750 0.000000000
#> 6      missense_point_mutations KRAS_Ex2_3       0.0250 0.000000000
#> 7            missense_point_mutations TET2       0.0775 0.007905694
#> 8            missense_point_mutations EZH2       0.1500 0.000000000
#> 9      missense_point_mutations CBL_Ex_8_9       0.1175 0.016873714
#> 10      missense_point_mutations IDH2_R140       0.0250 0.000000000
#> 11 missense_point_mutations SF3B1_Ex_12_15       0.0250 0.000000000
#> 12 missense_point_mutations JARID_2_Ex1_18       0.0450 0.015811388
#> 13     missense_point_mutations EED_Ex2_12       0.0500 0.000000000
#> 14          missense_point_mutations CEBPA       0.0250 0.000000000
#> 15          missense_point_mutations EPHB3       0.0275 0.007905694
#> 16          missense_point_mutations ETNK1       0.0250 0.000000000
#> 17          missense_point_mutations GATA2       0.0250 0.000000000
#> 18          missense_point_mutations IRAK4       0.0525 0.021889876
#> 19           missense_point_mutations MTA2       0.0250 0.000000000
#> 20          missense_point_mutations CSF3R       0.1000 0.000000000
#> 21            missense_point_mutations KIT       0.0550 0.019720266
#> 22                    nonsense_ins_del WT1       0.0500 0.000000000
#> 23                 nonsense_ins_del RUNX_1       0.0500 0.000000000
#> 24                  nonsense_ins_del CEBPA       0.0500 0.000000000
#> 25           nonsense_point_mutations TET2       0.1250 0.000000000
#> 26           nonsense_point_mutations EZH2       0.0250 0.000000000
#> 27          nonsense_point_mutations ASXL1       0.1000 0.000000000
#> 28          nonsense_point_mutations CSF3R       0.0750 0.000000000
#> 29                        Pattern XOR_EZH2       0.2000 0.000000000
#> 30                        Pattern OR_CSF3R       0.1500 0.000000000
#> 
#> $capri_aic
#>                                   SELECTED MEAN.PREDERR  SD.PREDERR
#> 1                             ins_del TET2       0.0775 0.007905694
#> 2                             ins_del EZH2       0.0250 0.000000000
#> 3                            ins_del ASXL1       0.1925 0.020581815
#> 4          missense_point_mutations SETBP1       0.3500 0.000000000
#> 5      missense_point_mutations NRAS_Ex2_3       0.0750 0.000000000
#> 6      missense_point_mutations KRAS_Ex2_3       0.0250 0.000000000
#> 7            missense_point_mutations TET2       0.0750 0.000000000
#> 8            missense_point_mutations EZH2       0.1500 0.000000000
#> 9      missense_point_mutations CBL_Ex_8_9       0.1450 0.019720266
#> 10      missense_point_mutations IDH2_R140       0.0250 0.000000000
#> 11 missense_point_mutations SF3B1_Ex_12_15       0.0250 0.000000000
#> 12 missense_point_mutations JARID_2_Ex1_18       0.0550 0.015811388
#> 13     missense_point_mutations EED_Ex2_12       0.0500 0.000000000
#> 14          missense_point_mutations CEBPA       0.0250 0.000000000
#> 15          missense_point_mutations EPHB3       0.0375 0.021245915
#> 16          missense_point_mutations ETNK1       0.0250 0.000000000
#> 17          missense_point_mutations GATA2       0.0250 0.000000000
#> 18          missense_point_mutations IRAK4       0.0475 0.018446620
#> 19           missense_point_mutations MTA2       0.0250 0.000000000
#> 20          missense_point_mutations CSF3R       0.1000 0.000000000
#> 21            missense_point_mutations KIT       0.0425 0.020581815
#> 22                    nonsense_ins_del WT1       0.0500 0.000000000
#> 23                 nonsense_ins_del RUNX_1       0.0500 0.000000000
#> 24                  nonsense_ins_del CEBPA       0.0425 0.012076147
#> 25           nonsense_point_mutations TET2       0.1250 0.000000000
#> 26           nonsense_point_mutations EZH2       0.0250 0.000000000
#> 27          nonsense_point_mutations ASXL1       0.1000 0.000000000
#> 28          nonsense_point_mutations CSF3R       0.0750 0.000000000
#> 29                        Pattern XOR_EZH2       0.2000 0.000000000
#> 30                        Pattern OR_CSF3R       0.1500 0.000000000
#> 

as.kfold.prederr(test_model_kfold, models='capri_aic')

#> $capri_aic
#>                                   SELECTED MEAN.PREDERR  SD.PREDERR
#> 1                             ins_del TET2       0.0875 0.021245915
#> 2                             ins_del EZH2       0.0250 0.000000000
#> 3                            ins_del ASXL1       0.1750 0.000000000
#> 4          missense_point_mutations SETBP1       0.3500 0.000000000
#> 5      missense_point_mutations NRAS_Ex2_3       0.0750 0.000000000
#> 6      missense_point_mutations KRAS_Ex2_3       0.0250 0.000000000
#> 7            missense_point_mutations TET2       0.0775 0.007905694
#> 8            missense_point_mutations EZH2       0.1500 0.000000000
#> 9      missense_point_mutations CBL_Ex_8_9       0.1175 0.016873714
#> 10      missense_point_mutations IDH2_R140       0.0250 0.000000000
#> 11 missense_point_mutations SF3B1_Ex_12_15       0.0250 0.000000000
#> 12 missense_point_mutations JARID_2_Ex1_18       0.0450 0.015811388
#> 13     missense_point_mutations EED_Ex2_12       0.0500 0.000000000
#> 14          missense_point_mutations CEBPA       0.0250 0.000000000
#> 15          missense_point_mutations EPHB3       0.0275 0.007905694
#> 16          missense_point_mutations ETNK1       0.0250 0.000000000
#> 17          missense_point_mutations GATA2       0.0250 0.000000000
#> 18          missense_point_mutations IRAK4       0.0525 0.021889876
#> 19           missense_point_mutations MTA2       0.0250 0.000000000
#> 20          missense_point_mutations CSF3R       0.1000 0.000000000
#> 21            missense_point_mutations KIT       0.0550 0.019720266
#> 22                    nonsense_ins_del WT1       0.0500 0.000000000
#> 23                 nonsense_ins_del RUNX_1       0.0500 0.000000000
#> 24                  nonsense_ins_del CEBPA       0.0500 0.000000000
#> 25           nonsense_point_mutations TET2       0.1250 0.000000000
#> 26           nonsense_point_mutations EZH2       0.0250 0.000000000
#> 27          nonsense_point_mutations ASXL1       0.1000 0.000000000
#> 28          nonsense_point_mutations CSF3R       0.0750 0.000000000
#> 29                        Pattern XOR_EZH2       0.2000 0.000000000
#> 30                        Pattern OR_CSF3R       0.1500 0.000000000
#>