as.selective.advantage.relations

Returns a dataframe with all the selective advantage relations in a TRONCO model. Confidence is also shown - see as.confidence. It is possible to specify a subset of events or models if multiple reconstruction have been performed.

as.selective.advantage.relations(
  x,
  events = as.events(x),
  models = names(x$model),
  type = "fit"
)

Arguments

x
A TRONCO model.

events
A subset of events as of as.events(x), all by default.

models
A subset of reconstructed models, all by default.

type
Either Prima Facie ('pf') or fit ('fit') probabilities, 'fit' by default.

Value

All the selective advantage relations in a TRONCO model

Examples

data(test_model)
as.selective.advantage.relations(test_model)

#> $capri_bic
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2            missense_point_mutations KIT
#> 2                        ins_del ASXL1           missense_point_mutations TET2
#> 3      missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 4  missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 5  missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 6        missense_point_mutations TET2                            ins_del EZH2
#> 7        missense_point_mutations EZH2                            ins_del TET2
#> 8        missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 9        missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 10        missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 11              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 12              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 13      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 14      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#>    OBS.SELECTS OBS.SELECTED TEMPORAL.PRIORITY PROBABILITY.RAISING
#> 1            3            2      1.421229e-06        5.460128e-22
#> 2            7            3      2.278343e-38        6.442772e-35
#> 3           14            4      1.426321e-55        7.635188e-63
#> 4            3            1      1.056706e-31        2.612641e-09
#> 5            3            1      2.827924e-29        1.055365e-20
#> 6            3            1      1.282066e-25        1.177708e-10
#> 7            6            3      9.373014e-28        7.840153e-28
#> 8            6            1      1.677847e-53        1.835656e-14
#> 9            6            1      8.126240e-53        2.348033e-21
#> 10           2            1      2.540051e-13        3.228035e-03
#> 11           2            1      4.511947e-12        1.966262e-07
#> 12           2            1      2.157029e-13        9.434522e-05
#> 13           4            3      1.650994e-08        2.926887e-32
#> 14           4            1      1.661066e-39        6.448439e-17
#>    HYPERGEOMETRIC
#> 1      0.00000000
#> 2      0.07368421
#> 3      0.00000000
#> 4      0.07500000
#> 5      0.00000000
#> 6      0.07500000
#> 7      0.05364372
#> 8      0.15000000
#> 9      0.00000000
#> 10     0.05000000
#> 11     0.00000000
#> 12     0.05000000
#> 13     0.02226721
#> 14     0.00000000
#> 
#> $capri_aic
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2     missense_point_mutations CBL_Ex_8_9
#> 2                         ins_del TET2            missense_point_mutations KIT
#> 3                        ins_del ASXL1     missense_point_mutations KRAS_Ex2_3
#> 4                        ins_del ASXL1           missense_point_mutations TET2
#> 5      missense_point_mutations SETBP1           nonsense_point_mutations EZH2
#> 6      missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 7  missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 8  missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 9        missense_point_mutations TET2                            ins_del EZH2
#> 10       missense_point_mutations TET2                  nonsense_ins_del CEBPA
#> 11       missense_point_mutations EZH2                            ins_del TET2
#> 12       missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 13       missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 14      missense_point_mutations CSF3R           missense_point_mutations TET2
#> 15        missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 16              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 17              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 18      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 19      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#> 20                    Pattern OR_CSF3R                  nonsense_ins_del CEBPA
#>    OBS.SELECTS OBS.SELECTED TEMPORAL.PRIORITY PROBABILITY.RAISING
#> 1            3            3      2.359743e-01        1.587286e-02
#> 2            3            2      1.421229e-06        5.460128e-22
#> 3            7            1      4.356112e-55        5.627097e-17
#> 4            7            3      2.278343e-38        6.442772e-35
#> 5           14            1      8.806722e-57        1.839890e-14
#> 6           14            4      1.426321e-55        7.635188e-63
#> 7            3            1      1.056706e-31        2.612641e-09
#> 8            3            1      2.827924e-29        1.055365e-20
#> 9            3            1      1.282066e-25        1.177708e-10
#> 10           3            2      5.985653e-07        1.056504e-05
#> 11           6            3      9.373014e-28        7.840153e-28
#> 12           6            1      1.677847e-53        1.835656e-14
#> 13           6            1      8.126240e-53        2.348033e-21
#> 14           4            3      5.618501e-11        6.632068e-05
#> 15           2            1      2.540051e-13        3.228035e-03
#> 16           2            1      4.511947e-12        1.966262e-07
#> 17           2            1      2.157029e-13        9.434522e-05
#> 18           4            3      1.650994e-08        2.926887e-32
#> 19           4            1      1.661066e-39        6.448439e-17
#> 20           6            2      7.398044e-44        2.822865e-06
#>    HYPERGEOMETRIC
#> 1      0.21356275
#> 2      0.00000000
#> 3      0.00000000
#> 4      0.07368421
#> 5      0.35000000
#> 6      0.00000000
#> 7      0.07500000
#> 8      0.00000000
#> 9      0.07500000
#> 10     0.14615385
#> 11     0.05364372
#> 12     0.15000000
#> 13     0.00000000
#> 14     0.27732794
#> 15     0.05000000
#> 16     0.00000000
#> 17     0.05000000
#> 18     0.02226721
#> 19     0.00000000
#> 20     0.28076923
#> 

as.selective.advantage.relations(test_model, events=as.events(test_model)[5:15,])

#> $capri_bic
#>                               SELECTS                                SELECTED
#> 1 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 2       missense_point_mutations EZH2          missense_point_mutations CEBPA
#>   OBS.SELECTS OBS.SELECTED TEMPORAL.PRIORITY PROBABILITY.RAISING HYPERGEOMETRIC
#> 1           3            1      1.056706e-31        2.612641e-09          0.075
#> 2           6            1      1.677847e-53        1.835656e-14          0.150
#> 
#> $capri_aic
#>                               SELECTS                                SELECTED
#> 1 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 2       missense_point_mutations EZH2          missense_point_mutations CEBPA
#>   OBS.SELECTS OBS.SELECTED TEMPORAL.PRIORITY PROBABILITY.RAISING HYPERGEOMETRIC
#> 1           3            1      1.056706e-31        2.612641e-09          0.075
#> 2           6            1      1.677847e-53        1.835656e-14          0.150
#> 

as.selective.advantage.relations(test_model, events=as.events(test_model)[5:15,], type='pf')

#> $pf
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2     missense_point_mutations CBL_Ex_8_9
#> 2                         ins_del TET2            missense_point_mutations KIT
#> 3                         ins_del TET2                 nonsense_ins_del RUNX_1
#> 4                        ins_del ASXL1     missense_point_mutations KRAS_Ex2_3
#> 5                        ins_del ASXL1           missense_point_mutations TET2
#> 6                        ins_del ASXL1     missense_point_mutations CBL_Ex_8_9
#> 7      missense_point_mutations SETBP1     missense_point_mutations NRAS_Ex2_3
#> 8      missense_point_mutations SETBP1     missense_point_mutations CBL_Ex_8_9
#> 9      missense_point_mutations SETBP1 missense_point_mutations JARID_2_Ex1_18
#> 10     missense_point_mutations SETBP1          missense_point_mutations EPHB3
#> 11     missense_point_mutations SETBP1          missense_point_mutations CSF3R
#> 12     missense_point_mutations SETBP1           nonsense_point_mutations TET2
#> 13     missense_point_mutations SETBP1           nonsense_point_mutations EZH2
#> 14     missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 15     missense_point_mutations SETBP1          nonsense_point_mutations CSF3R
#> 16     missense_point_mutations SETBP1                        Pattern OR_CSF3R
#> 17 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 18 missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 19       missense_point_mutations TET2                            ins_del EZH2
#> 20       missense_point_mutations TET2                    nonsense_ins_del WT1
#> 21       missense_point_mutations TET2                  nonsense_ins_del CEBPA
#> 22       missense_point_mutations EZH2                            ins_del TET2
#> 23       missense_point_mutations EZH2     missense_point_mutations NRAS_Ex2_3
#> 24       missense_point_mutations EZH2     missense_point_mutations CBL_Ex_8_9
#> 25       missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 26       missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 27       missense_point_mutations EZH2          missense_point_mutations IRAK4
#> 28       missense_point_mutations EZH2            missense_point_mutations KIT
#> 29      missense_point_mutations CSF3R           missense_point_mutations TET2
#> 30      missense_point_mutations CSF3R                    nonsense_ins_del WT1
#> 31      missense_point_mutations CSF3R                  nonsense_ins_del CEBPA
#> 32      missense_point_mutations CSF3R          nonsense_point_mutations CSF3R
#> 33        missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 34              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 35              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 36       nonsense_point_mutations TET2     missense_point_mutations NRAS_Ex2_3
#> 37       nonsense_point_mutations TET2 missense_point_mutations JARID_2_Ex1_18
#> 38       nonsense_point_mutations TET2          missense_point_mutations CSF3R
#> 39      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 40      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#> 41                    Pattern XOR_EZH2                            ins_del TET2
#> 42                    Pattern XOR_EZH2     missense_point_mutations NRAS_Ex2_3
#> 43                    Pattern XOR_EZH2          missense_point_mutations CEBPA
#> 44                    Pattern XOR_EZH2          missense_point_mutations ETNK1
#> 45                    Pattern XOR_EZH2          missense_point_mutations IRAK4
#> 46                    Pattern XOR_EZH2            missense_point_mutations KIT
#> 47                    Pattern OR_CSF3R           missense_point_mutations TET2
#> 48                    Pattern OR_CSF3R                    nonsense_ins_del WT1
#> 49                    Pattern OR_CSF3R                  nonsense_ins_del CEBPA
#>    OBS.SELECTS OBS.SELECTED TEMPORAL.PRIORITY PROBABILITY.RAISING
#> 1            3            3      2.359743e-01        1.587286e-02
#> 2            3            2      1.421229e-06        5.460128e-22
#> 3            3            1      3.262925e-27        5.352680e-07
#> 4            7            1      4.356112e-55        5.627097e-17
#> 5            7            3      2.278343e-38        6.442772e-35
#> 6            7            3      4.899826e-40        4.004727e-02
#> 7           14            3      4.357290e-56        7.755853e-24
#> 8           14            3      4.145089e-56        7.691348e-28
#> 9           14            1      9.175701e-57        5.311779e-11
#> 10          14            1      1.287253e-56        2.607680e-20
#> 11          14            4      6.828657e-56        5.697137e-19
#> 12          14            5      5.185189e-55        3.729102e-03
#> 13          14            1      8.806722e-57        1.839890e-14
#> 14          14            4      1.426321e-55        7.635188e-63
#> 15          14            3      3.972375e-56        3.011047e-32
#> 16          14            6      5.477097e-53        8.260024e-23
#> 17           3            1      1.056706e-31        2.612641e-09
#> 18           3            1      2.827924e-29        1.055365e-20
#> 19           3            1      1.282066e-25        1.177708e-10
#> 20           3            1      1.806902e-26        3.254703e-10
#> 21           3            2      5.985653e-07        1.056504e-05
#> 22           6            3      9.373014e-28        7.840153e-28
#> 23           6            3      3.120373e-25        8.401800e-13
#> 24           6            3      3.690851e-30        2.969062e-02
#> 25           6            1      1.677847e-53        1.835656e-14
#> 26           6            1      8.126240e-53        2.348033e-21
#> 27           6            1      3.167398e-52        3.631082e-27
#> 28           6            2      5.489839e-41        8.185249e-09
#> 29           4            3      5.618501e-11        6.632068e-05
#> 30           4            1      3.309532e-45        1.348268e-12
#> 31           4            2      1.860213e-26        3.703793e-06
#> 32           4            3      9.416075e-09        6.601609e-13
#> 33           2            1      2.540051e-13        3.228035e-03
#> 34           2            1      4.511947e-12        1.966262e-07
#> 35           2            1      2.157029e-13        9.434522e-05
#> 36           5            3      9.388325e-16        8.545585e-03
#> 37           5            1      4.217177e-49        3.823050e-10
#> 38           5            4      5.771054e-04        1.704655e-06
#> 39           4            3      1.650994e-08        2.926887e-32
#> 40           4            1      1.661066e-39        6.448439e-17
#> 41           8            3      1.657722e-44        8.907302e-28
#> 42           8            3      1.100280e-43        3.527613e-10
#> 43           8            1      2.432208e-56        1.836811e-14
#> 44           8            1      3.987969e-56        1.051434e-21
#> 45           8            1      4.919078e-56        1.589068e-27
#> 46           8            2      4.242942e-52        8.546011e-09
#> 47           6            3      4.947292e-32        2.296749e-04
#> 48           6            1      1.058573e-52        1.598725e-13
#> 49           6            2      7.398044e-44        2.822865e-06
#>    HYPERGEOMETRIC
#> 1      0.21356275
#> 2      0.00000000
#> 3      0.07500000
#> 4      0.00000000
#> 5      0.07368421
#> 6      0.44777328
#> 7      0.03684211
#> 8      0.27631579
#> 9      0.35000000
#> 10     0.00000000
#> 11     0.11450925
#> 12     0.22238028
#> 13     0.35000000
#> 14     0.00000000
#> 15     0.03684211
#> 16     0.09909910
#> 17     0.07500000
#> 18     0.00000000
#> 19     0.07500000
#> 20     0.07500000
#> 21     0.14615385
#> 22     0.05364372
#> 23     0.05364372
#> 24     0.39433198
#> 25     0.15000000
#> 26     0.00000000
#> 27     0.00000000
#> 28     0.01923077
#> 29     0.27732794
#> 30     0.10000000
#> 31     0.19230769
#> 32     0.02226721
#> 33     0.05000000
#> 34     0.00000000
#> 35     0.05000000
#> 36     0.33755061
#> 37     0.12500000
#> 38     0.06899004
#> 39     0.02226721
#> 40     0.00000000
#> 41     0.09635628
#> 42     0.09635628
#> 43     0.20000000
#> 44     0.00000000
#> 45     0.00000000
#> 46     0.03589744
#> 47     0.39433198
#> 48     0.15000000
#> 49     0.28076923
#>