hypothesis.add.homologous

Add all the hypotheses related to homologou events

hypothesis.add.homologous(
  x,
  pattern.cause = "*",
  pattern.effect = "*",
  genes = as.genes(x),
  silent = FALSE
)

Arguments

x

A TRONCO compliant dataset.

pattern.cause

Possibile causes for the pattern.

pattern.effect

Possibile effects for the pattern.

genes

List of genes to be considered as possible homologous. For these genes, all the types of mutations will be considered functionally equivalent.

silent

A parameter to disable/enable verbose messages.

Value

A TRONCO compliant object with the added hypotheses