import.genotypes

Import a matrix of 0/1 alterations as a TRONCO compliant dataset. Input "geno" can be either a dataframe or a file name. In any case the dataframe or the table stored in the file must have a column for each altered gene and a rows for each sample. Colnames will be used to determine gene names, if data is loaded from file the first column will be assigned as rownames. For details and examples regarding the loading functions provided by the package we refer to the Vignette Section 3.

import.genotypes(geno, event.type = "variant", color = "Darkgreen")

Arguments

geno

Either a dataframe or a filename

event.type

Any 1 in "geno" will be interpreted as a an observed alteration labeled with type "event.type"

color

This is the color used for visualization of events labeled as of "event.type"

Value

A TRONCO compliant dataset