oncoprint.cbio

export input for cbio visualization at http://www.cbioportal.org/public-portal/oncoprinter.jsp

oncoprint.cbio(
  x,
  file = "oncoprint-cbio.txt",
  hom.del = "Homozygous Loss",
  het.loss = "Heterozygous Loss",
  gain = "Low-level Gain",
  amp = "High-level Gain"
)

Arguments

x
A TRONCO compliant dataset.

file
name of the file where to save the output

hom.del
type of Homozygous Deletion

het.loss
type of Heterozygous Loss

gain
type of Gain

amp
type of Amplification

Value

A file containing instruction for the CBio visualization Tool

Examples

data(crc_gistic)
gistic = import.GISTIC(crc_gistic)

#> *** Using full GISTIC: #dim  9  x  6 
#> *** GISTIC input format conversion started.
#> Converting input data to character for import speedup.
#> Creating  24 events for 6 genes 
#> Extracting "Homozygous Loss" events (GISTIC = -2) 
#> Extracting "Heterozygous Loss" events (GISTIC = -1) 
#> Extracting "Low-level Gain" events (GISTIC = +1) 
#> Extracting "High-level Gain" events (GISTIC = +2) 
#> Transforming events in TRONCO data types ..... 
#> *** Binding events for 4 datasets.
#> *** Data extracted, returning only events observed in at least one sample 
#>  Number of events: n = 7 
#>  Number of genes: |G| = 6 
#>  Number of samples: m = 9 

oncoprint.cbio(gistic)