Tronco Chow Liu

Reconstruct a progression model using Chow Liu algorithm combined with probabilistic causation. For details and examples regarding the inference process and on the algorithm implemented in the package, we refer to the Vignette Section 6.

tronco.chowliu(
  data,
  regularization = c("bic", "aic"),
  do.boot = TRUE,
  nboot = 100,
  pvalue = 0.05,
  min.boot = 3,
  min.stat = TRUE,
  boot.seed = NULL,
  silent = FALSE,
  epos = 0,
  eneg = 0
)

Arguments

data

A TRONCO compliant dataset.

regularization

Select the regularization for the likelihood estimation, e.g., BIC, AIC.

do.boot

A parameter to disable/enable the estimation of the error rates give the reconstructed model.

nboot

Number of bootstrap sampling (with rejection) to be performed when estimating the selective advantage scores.

pvalue

Pvalue to accept/reject the valid selective advantage relations.

min.boot

Minimum number of bootstrap sampling to be performed.

min.stat

A parameter to disable/enable the minimum number of bootstrap sampling required besides nboot if any sampling is rejected.

boot.seed

Initial seed for the bootstrap random sampling.

silent

A parameter to disable/enable verbose messages.

epos

Error rate of false positive errors.

eneg

Error rate of false negative errors.

Value

A TRONCO compliant object with reconstructed model

Examples

data(test_dataset_no_hypos)
recon = tronco.chowliu(test_dataset_no_hypos, nboot = 1)
#> *** Checking input events.
#> *** Inferring a progression model with the following settings.
#> 	Dataset size: n = 40, m = 28.
#> 	Algorithm: Chow Liu with "bic, aic" regularization	Random seed: NULL.
#> 	Bootstrap iterations (Wilcoxon): 1.
#> 		exhaustive bootstrap: TRUE.
#> 		p-value: 0.05.
#> 		minimum bootstrapped scores: 3.
#> *** Bootstraping selective advantage scores (prima facie).
#> 	........
#> 	Evaluating "temporal priority" (Wilcoxon, p-value 0.05)
#> 	Evaluating "probability raising" (Wilcoxon, p-value 0.05)
#> *** Loop detection found loops to break.
#> 	Removed 3 edges out of 38 (8%)
#> *** Performing likelihood-fit with regularization bic .
#> *** Performing likelihood-fit with regularization aic .
#> *** Evaluating BIC / AIC / LogLik informations.
#> The reconstruction has been successfully completed in 00h:00m:01s