as.kfold.posterr

Returns a dataframe with all the posterior classification error score in a TRONCO model. It is possible to specify a subset of events or models if multiple reconstruction have been performed.

as.kfold.posterr(
  x,
  events = as.events(x),
  models = names(x$model),
  values = FALSE,
  table = FALSE
)

Arguments

x
A TRONCO model.

events
A subset of events as of as.events(x), all by default.

models
A subset of reconstructed models, all by default.

values
If you want to see also the values

table
Keep the original table (defaul false)

Value

All the posterior classification error scores in a TRONCO model

Examples

data(test_model_kfold)
data(test_model)
as.kfold.posterr(test_model_kfold)
#> $capri_bic
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2            missense_point_mutations KIT
#> 2                        ins_del ASXL1           missense_point_mutations TET2
#> 3      missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 4  missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 5  missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 6        missense_point_mutations EZH2                            ins_del TET2
#> 7        missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 8        missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 9         missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 10              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 11              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 12      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 13      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#>    MEAN.POSTERR  SD.POSTERR
#> 1        0.0475 0.021889876
#> 2        0.0800 0.015811388
#> 3        0.1000 0.000000000
#> 4        0.0525 0.018446620
#> 5        0.0525 0.007905694
#> 6        0.0875 0.021245915
#> 7        0.0250 0.000000000
#> 8        0.0250 0.000000000
#> 9        0.0500 0.000000000
#> 10       0.0475 0.007905694
#> 11       0.0500 0.000000000
#> 12       0.1225 0.007905694
#> 13       0.0325 0.012076147
#> 
#> $capri_aic
#>                                SELECTS                                SELECTED
#> 1                         ins_del TET2     missense_point_mutations CBL_Ex_8_9
#> 2                         ins_del TET2            missense_point_mutations KIT
#> 3                        ins_del ASXL1     missense_point_mutations KRAS_Ex2_3
#> 4                        ins_del ASXL1           missense_point_mutations TET2
#> 5      missense_point_mutations SETBP1           nonsense_point_mutations EZH2
#> 6      missense_point_mutations SETBP1          nonsense_point_mutations ASXL1
#> 7  missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 8  missense_point_mutations NRAS_Ex2_3          missense_point_mutations IRAK4
#> 9        missense_point_mutations TET2                  nonsense_ins_del CEBPA
#> 10       missense_point_mutations EZH2                            ins_del TET2
#> 11       missense_point_mutations EZH2          missense_point_mutations CEBPA
#> 12       missense_point_mutations EZH2          missense_point_mutations ETNK1
#> 13      missense_point_mutations CSF3R           missense_point_mutations TET2
#> 14        missense_point_mutations KIT                 nonsense_ins_del RUNX_1
#> 15              nonsense_ins_del CEBPA     missense_point_mutations EED_Ex2_12
#> 16              nonsense_ins_del CEBPA                    nonsense_ins_del WT1
#> 17      nonsense_point_mutations ASXL1     missense_point_mutations CBL_Ex_8_9
#> 18      nonsense_point_mutations ASXL1          missense_point_mutations EPHB3
#> 19                    Pattern XOR_EZH2                           ins_del ASXL1
#> 20                    Pattern OR_CSF3R                  nonsense_ins_del CEBPA
#>    MEAN.POSTERR  SD.POSTERR
#> 1        0.0875 0.017677670
#> 2        0.0475 0.018446620
#> 3        0.0250 0.000000000
#> 4        0.0750 0.000000000
#> 5        0.0250 0.000000000
#> 6        0.1000 0.000000000
#> 7        0.0450 0.010540926
#> 8        0.0475 0.021889876
#> 9        0.0500 0.000000000
#> 10       0.0825 0.016873714
#> 11       0.0250 0.000000000
#> 12       0.0250 0.000000000
#> 13       0.0775 0.007905694
#> 14       0.0500 0.000000000
#> 15       0.0500 0.000000000
#> 16       0.0500 0.000000000
#> 17       0.1250 0.000000000
#> 18       0.0300 0.010540926
#> 19       0.1900 0.033747428
#> 20       0.0500 0.000000000
#> 
as.kfold.posterr(test_model_kfold, events=as.events(test_model)[5:15,])
#> $capri_bic
#>                               SELECTS                                SELECTED
#> 1 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 2       missense_point_mutations EZH2          missense_point_mutations CEBPA
#>   MEAN.POSTERR SD.POSTERR
#> 1       0.0525 0.01844662
#> 2       0.0250 0.00000000
#> 
#> $capri_aic
#>                               SELECTS                                SELECTED
#> 1 missense_point_mutations NRAS_Ex2_3 missense_point_mutations JARID_2_Ex1_18
#> 2       missense_point_mutations EZH2          missense_point_mutations CEBPA
#>   MEAN.POSTERR SD.POSTERR
#> 1        0.045 0.01054093
#> 2        0.025 0.00000000
#>